Rename file pix-link.vue to PixLink.vue
