+++ components/PixLink.vue
