+++ plugins/components.js
@@ -1,4 +1,4 @@
 import Vue from 'vue'
-import PixLink from '~/components/pix-link'
+import PixLink from '@/components/PixLink'
 
 Vue.component('PixLink', PixLink)
